commit_project

+++ Scripts/jsShopping.js
@@ -8,6 +8,7 @@
         if (tQuantity != '') {
             quatity = parseInt(tQuantity);
         }
+
         //alert(id + " " + quatity);
         $.ajax({
             url: '/shoppingcart/addtocart',
@@ -16,55 +17,63 @@
             success: function (rs) {
                 if (rs.Success) {
                     $('#checkout_items').html(rs.Count);
+                    alert(rs.msg);
                 }
             }
         });
     });
-
-    $('body').on('click', '.btnDelete', function (e) {
-        e.preventDefault();
-        var id = $(this).data('id');
-        $.ajax({
-            url: '/shoppingcart/Delete',
-            type: 'POST',
-            data: { id: id },
-            success: function (rs) {
-                if (rs.Success) {
-                    $('#checkout_items').html(rs.Count);
-                    $('#trow_' + id).remove();
-                }
-            }
-        });
-        LoadCart();
-    });
-
-    $('body').on('click', '.btnDeleteAll', function (e) {
-        e.preventDefault();
-        DeleteAll();
-    });
-
     $('body').on('click', '.btnUpdate', function (e) {
         e.preventDefault();
         var id = $(this).data("id");
         var quantity = $('#Quantity_' + id).val();
         Update(id, quantity);
+
+    });
+    $('body').on('click', '.btnDeleteAll', function (e) {
+        e.preventDefault(); 
+        var conf = confirm('Bạn muốn xóa hết sản phẩm trong giỏ hàng?'); 
+        //debugger; 
+        if (conf == true) { 
+            DeleteAll(); 
+        } 
+    });
+
+    $('body').on('click', '.btnDelete', function (e) {
+        e.preventDefault(); 
+        var id = $(this).data('id'); 
+        var conf = confirm('Bạn muốn xóa sản phẩm này khỏi giỏ hàng?'); 
+        if (conf == true) { 
+            $.ajax({
+                url: '/shoppingcart/Delete',
+                type: 'POST',
+                data: { id: id },
+                success: function (rs) {
+                    if (rs.Success) {
+                        $('#checkout_items').html(rs.Count);
+                        $('#trow_' + id).remove();
+                        LoadCart();
+                    }
+                }
+            });
+        }
+
     });
 });
 
 
+
 function ShowCount() {
     $.ajax({
-        url: '/ShoppingCart/ShowCount',
+        url: '/shoppingcart/ShowCount',
         type: 'GET',
         success: function (rs) {
             $('#checkout_items').html(rs.Count);
         }
     });
 }
-
 function DeleteAll() {
     $.ajax({
-        url: '/ShoppingCart/DeleteAll',
+        url: '/shoppingcart/DeleteAll',
         type: 'POST',
         success: function (rs) {
             if (rs.Success) {
@@ -73,12 +82,11 @@ function DeleteAll() {
         }
     });
 }
-
 function Update(id, quantity) {
     $.ajax({
-        url: '/ShoppingCart/Update',
-        data: { id: id, quantity: quantity },
+        url: '/shoppingcart/Update',
         type: 'POST',
+        data: { id: id, quantity: quantity },
         success: function (rs) {
             if (rs.Success) {
                 LoadCart();
@@ -89,7 +97,7 @@ function Update(id, quantity) {
 
 function LoadCart() {
     $.ajax({
-        url: '/ShoppingCart/Partial_Item_Cart',
+        url: '/shoppingcart/Partial_Item_Cart',
         type: 'GET',
         success: function (rs) {
             $('#load_data').html(rs);
